# docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: 新点 SaaS 造价系统 API
  version: 1.0.0-current
  summary: 当前代码已注册的 V1 API 最小契约
  description: |
    自动生成文件，请勿手工编辑。
    生成命令：npm run docs:openapi-current
servers:
  - url: http://localhost:3000
    description: Local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    ObjectEnvelope:
      type: object
      additionalProperties: true
    ListEnvelope:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            additionalProperties: true
      additionalProperties: true
paths:
...
  /v1/engine/calculate:
    post:
      operationId: postEngineCalculate
      summary: POST /v1/engine/calculate
      tags:
        - Pricing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/fee-templates:
    get:
      operationId: getFeeTemplates
      summary: GET /v1/fee-templates
      tags:
        - Pricing
      security:
        - bearerAuth: []
      parameters:
      - name: regionCode
        in: query
        description: Region code.
        required: false
        schema:
          type: string
      - name: stageCode
        in: query
        description: Project stage code.
        required: false
        schema:
          type: string
      - name: activeOnly
        in: query
        description: Only return active records when true.
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/fee-templates/{feeTemplateId}:
    get:
      operationId: getFeeTemplatesByFeeTemplateId
      summary: GET /v1/fee-templates/:feeTemplateId
      tags:
        - Pricing
      security:
        - bearerAuth: []
      parameters:
      - name: feeTemplateId
        in: path
        description: feeTemplateId path parameter.
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/price-versions:
    get:
      operationId: getPriceVersions
      summary: GET /v1/price-versions
      tags:
        - Pricing
      security:
        - bearerAuth: []
      parameters:
      - name: regionCode
        in: query
        description: Region code.
        required: false
        schema:
          type: string
      - name: disciplineCode
        in: query
        description: Project discipline code.
        required: false
        schema:
          type: string
      - name: activeOnly
        in: query
        description: Only return active records when true.
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/price-versions/{priceVersionId}/items:
    get:
      operationId: getPriceVersionsByPriceVersionIdItems
      summary: GET /v1/price-versions/:priceVersionId/items
      tags:
        - Pricing
      security:
        - bearerAuth: []
      parameters:
      - name: priceVersionId
        in: path
        description: priceVersionId path parameter.
        required: true
        schema:
          type: string
      - name: quotaCode
        in: query
        description: quotaCode query parameter.
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...